Protologging: Add protocol buffer logging.
Instead of sending events to ASL they're written to delimited protocol
buffer messages in a single file, which is then rotated after a certain
number of events of when a SIGHUP is received. This generally uses less
CPU as it requires a lot less string processing and allows for more data
to be cleanly kept in the logs.